services/gesture-service/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gesture-service"
version = "0.1.0"
description = "Gesture detector and MQTT publisher for smart mirror"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "mediapipe>=0.10,<0.11",
  "opencv-python-headless>=4.10,<5.0",
  "paho-mqtt>=2.1,<3.0",
  "pydantic>=2.8,<3.0",
  "pyyaml>=6.0,<7.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.7,<1.0"
]

[project.scripts]
gesture-service = "gesture_service.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: tests that require service-level wiring"
]

[tool.ruff]
line-length = 100
target-version = "py311"
